Add quickfix support to vscode-lang-tidy
This patch adds quickfix support so that the many
clang-tidy checks with fixes can automatically applies
to the code.

Unfortunatelly, the patch also contains some refactoring,
which helped me to find my way around the code.
If there is general interest in merging this patch, I can, of course, remove these
refactorings.